deployment/update-db/requirements.txt
pandas
psycopg2-binary
geopy
locationtagger
spacy
selenium
webdriver-manager